Add support for ontology detail view
The database was extended to include relationship between OntologyTerm instances and the relationships were created using the odo files.

A new url endpoints, views and HTML templates for displaying ontologies and the related phenotypes.
Global search also supports searching for ontologies and the search result will link to the above mentioned ontology detail view.
This fixes #11