Use postgres DB for production deployment
For production deployment spin up a separate postgres 11 container that
hosts the data. For local development the sqlite database is continued
to be used.
The sqlite db is added to .gitignore so it is not updated in future.
Refactor the Dockerfile to create an immutable container instead of
mounting the code as volume (this is necessary for th CI/CD).
Also make sure to collect the static asssets and copy them to the nginx
web root